factories/base.py: check if bin actually has ghostpad before attempting to remove it

+++ pitivi/factories/base.py
@@ -461,11 +461,12 @@ class SourceFactory(ObjectFactory):
         self.debug("dbin:%r, pad:%r, topbin:%r", dbin, pad, topbin)
 
         # work around for http://bugzilla.gnome.org/show_bug.cgi?id=590735
-        die = gst.Pad("die", gst.PAD_SRC)
-        topbin.ghostpad.set_target(die)
+        if hasattr(topbin, "ghostpad"):
+            die = gst.Pad("die", gst.PAD_SRC)
+            topbin.ghostpad.set_target(die)
 
-        topbin.remove_pad(topbin.ghostpad)
-        del topbin.ghostpad
+            topbin.remove_pad(topbin.ghostpad)
+            del topbin.ghostpad
 
         if hasattr(topbin, "volume"):
             pad.unlink(topbin.aconv.get_pad("sink"))
